Add files via upload
This project is to demonstrate (OOP) Object-Oriented programming principles using Java Interfaces and classes. It involves creating different types of meal classes that implement Meal_Cook and display information about the meals.

This project instructs you to write four different classes in separate source code files. The first-class name BackyardBBQ shows you’ve either used propane or charcoal to grill. The 2nd class is named ThanksgivingDinner sharing that you can use the oven and a roasting pan for the turkey. The third class is named HeartyStew and shows whether you used a crockpot or a stovetop to make the soup. The 4th class is called CookingDinner which includes your main method that runs the program effectively while also printing out the course details. Meal_Cook acts as the interface for BackyardBBQ, ThanksgivingDinner and HeartyStew.